+++ tests/loki_plugin/plugintest/client/type.go
@@ -15,7 +15,7 @@
 package client
 
 import (
-	"github.com/credativ/vali/pkg/promtail/api"
+	"github.com/credativ/vali/pkg/valitail/api"
 	"github.com/prometheus/common/model"
 )
